refactor: CodeConstant.tsを削除してslack.tsに寄せる
