+++ src/listeners/res_cem_edit.ts
@@ -1,7 +1,6 @@
 import { app } from '../initializers/bolt'
 import { FieldValue, firestore } from '../initializers/firebase'
-import { Challenge, Message, Project } from '../types/slack'
-import { CallbackId } from '../types/CodeConstant'
+import { CallbackId, Challenge, Message, Project } from '../types/slack'
 
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
 class ModalDto {
